Extract Constraints class
Extract a class from Validators::UriValidator.  Keep indentation
unchanged for now (therefore inconsistent a bit) so that Git tracks
history correctly.
