+++ lib/uri_format_validator.rb
@@ -2,6 +2,7 @@
 #
 
 require "uri_format_validator/version"
+require "uri_format_validator/constraints"
 require "uri_format_validator/localization"
 require "uri_format_validator/reacher"
 require "uri_format_validator/validators"
