+++ lib/uri_format_validator/validators/uri_validator.rb
@@ -13,28 +13,21 @@ module UriFormatValidator
     # TODO: documentation
     #
     class UriValidator < ::ActiveModel::EachValidator
+      attr_reader :constraints
+
       def initialize(options)
+        @constraints = Constraints.new(options)
         super(options)
       end
 
       def validate_each(record, attribute, value)
-        success = catch(STOP_VALIDATION) do
-          do_checks(value.to_s)
-          true
+        uri = string_to_uri(value)
+        unless uri && constraints.match?(uri)
+          set_failure_message(record, attribute)
         end
-        success || set_failure_message(record, attribute)
       end
 
       private
-
-      STOP_VALIDATION = Object.new.freeze
-
-      def do_checks(uri_string)
-        uri = string_to_uri(uri_string)
-        invalid unless uri
-
-        validate_against_options(uri, :scheme, :host, :retrievable)
-      end
 
       def string_to_uri(uri_string)
         Addressable::URI.parse(uri_string)
@@ -44,31 +37,6 @@ module UriFormatValidator
 
       def set_failure_message(record, attribute)
         record.errors[attribute] << failure_message
-      end
-
-      def invalid
-        throw STOP_VALIDATION
-      end
-
-      def validate_against_options(uri, *option_keys_list)
-        option_keys_list.each do |option_name|
-          next unless options[option_name]
-          send(:"validate_#{option_name}", options[option_name], uri)
-        end
-      end
-
-      def validate_host(host_or_hosts, uri)
-        hosts = Array.wrap(host_or_hosts)
-        invalid unless hosts.any? { |h| h === uri.hostname }
-      end
-
-      def validate_scheme(scheme_or_schemes, uri)
-        schemes = Array.wrap(scheme_or_schemes)
-        invalid unless schemes.any? { |s| s === uri.scheme }
-      end
-
-      def validate_retrievable(option, uri)
-        invalid if option && !Reacher.new(uri).retrievable?
       end
 
       def failure_message
